readme.md: Fix (some) broken links
Let's fix the links that are not longer valid. When possible, update
with the new link; otherwise, remove the associated entry.

After this change, the link checker still stumbles on a few links, but
they are valid in the browser, and I see no obvious way to fix them. I
suspect some CDNs reject bots or automated link checks and we're hitting
this.

Signed-off-by: Quentin Monnet <qmo@qmon.net>